moved the sanitizing of the code for preview to the child
